requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
websockets==15.0.1
python-dotenv==1.2.1
openai==2.8.1
sounddevice==0.5.3
scipy==1.16.3
soundfile==0.13.1
pytest==9.0.1
google-genai
pytubefix